backend/requirements.txt
cchardet
coverage
codecov
cerberus
Flask
Flask-API
Flask-Migrate
Flask-Script 
flask-bootstrap
flask-marshmallow
flask-restplus
flask-cors
flask-swagger
flasgger
itsdangerous==0.24
Jinja2
kombu
pytz
werkzeug
python-dateutil==2.4.2
python-dotenv
psycopg2
psycopg2-binary
gunicorn==20.0.4
requests
six==1.9.0
SQLAlchemy
sqlalchemy-migrate
marshmallow-sqlalchemy
